[feat] CI - Only generate schema on directiroy that is updated

+++ shared/tools/composer/generate_connectors_config_json_schemas.sh
@@ -68,18 +68,22 @@ connector_directories_path=$(find . -type d -name "$CONNECTOR_METADATA_DIRECTORY
 for connector_directory_path in $connector_directories_path
 do
   if [ -d "$connector_directory_path" ]; then
-    echo "> Looking for a config loader in " "$connector_directory_path"
-    requirements_file=$(find_requirements_txt "$connector_directory_path")
-    echo "Found requirements.txt: " "$requirements_file"
-    if [ -f "$requirements_file" ] && grep -q "pydantic-settings" "$requirements_file"; then
-      (
-        activate_venv "$connector_directory_path"
-        generator_path=$(find . -name "generate_connectors_config_json_schemas.py.sample")
-        cp "$generator_path" "$connector_directory_path/generate_connectors_config_json_schemas_tmp.py"
-        python "$connector_directory_path/generate_connectors_config_json_schemas_tmp.py"
-        rm "$connector_directory_path/generate_connectors_config_json_schemas_tmp.py"
-        deactivate_venv "$connector_directory_path/$VENV_NAME"
-      ) &
+    # Only generate schema for directory that changed
+    directory_has_changed=$(git diff "$connector_directory_path")
+    if [ -n "$directory_has_changed" ]; then
+      echo "> Looking for a config loader in " "$connector_directory_path"
+      requirements_file=$(find_requirements_txt "$connector_directory_path")
+      echo "Found requirements.txt: " "$requirements_file"
+      if [ -f "$requirements_file" ] && grep -q "pydantic-settings" "$requirements_file"; then
+        (
+          activate_venv "$connector_directory_path"
+          generator_path=$(find . -name "generate_connectors_config_json_schemas.py.sample")
+          cp "$generator_path" "$connector_directory_path/generate_connectors_config_json_schemas_tmp.py"
+          python "$connector_directory_path/generate_connectors_config_json_schemas_tmp.py"
+          rm "$connector_directory_path/generate_connectors_config_json_schemas_tmp.py"
+          deactivate_venv "$connector_directory_path/$VENV_NAME"
+        ) &
+      fi
     fi
   fi
 done
